fix: linting, move icon into UIComponentsLib, add missing jdoc
