+++ packages/webapp/src/webview/utilities/DeepMerge.ts
@@ -1,9 +1,10 @@
 /**
- * Performs deep merge on two objects.
- * Parameters are not mutated.
+ * Performs deep merge on two objects. Parameters are not mutated.
  * Arrays are copied as references and are not merged.
- * @param a
- * @param b
+ *
+ * @param a - object one
+ * @param b - object two
+ * @returns - merged object
  */
 export function deepMerge(a: any, b: any): Record<string, unknown> {
     const result = Object.keys(a).length ? deepMerge({}, a) : {};
@@ -23,6 +24,12 @@ export function deepMerge(a: any, b: any): Record<string, unknown> {
     return result;
 }
 
+/**
+ * Checks if passed item is object.
+ *
+ * @param item - item to check
+ * @returns - true: is object; false: no object
+ */
 function isObject(item: unknown): boolean {
     if (!item) {
         return false;
